// 18-06-2025/user-dashboard/src/app/core/models/user.model.ts
export interface User {
  id?: number;
  firstName: string;
  lastName: string;
  age: number;
  gender: string;
  email: string;
  phone: string;
  username: string;
  password: string;
  birthDate: string;
  image: string;
  bloodGroup: string;
  height: number;
  weight: number;
  eyeColor: string;
  hair: {
    color: string;
    type: string;
  };
  ip: string;
  address: {
    address: string;
    city: string;
    state: string;
    stateCode: string;
    postalCode: string;
    coordinates: {
      lat: number;
      lng: number;
    };
  };
  role?: string;
}


export interface UserFilters{
  gender?: string;
  role?: string;
  state?: string;
  minAge?: number;
  maxAge?: number;
}